Add additional file for custom decoder tests
