+++ Tests/JetworkingTests/MockData/HeaderFields.swift
@@ -1,0 +1,9 @@
+import Foundation
+@testable import Jetworking
+
+enum HeaderFields {
+    static let additional: [String: String] = [
+        "Accept": "application/json",
+        "Content-Type": "application/json"
+    ]
+}
